Added camera angles
The reference images are just where I want the "animatronics" to be, I recommend just coding them into their positions when we're coding the camera angles. Some angles might have to be cropped a bit but I think it'll be fine. The camera disabled is for both the left and right storage rooms and that's where we can go crazy with the sfx. Also I want the blue emoji guy when he's in the vents hallway A, or backstage (but only when someone has moved) to have randomized faces each time you open his camera. When he's backstage and someone hasn't moved yet, he should have the same face as the reference image