Added Scratch Reference (what I have so far)
A Scratch reference is a project coded in the Scratch coding language and has nothing to do with the final C# code, it's just a reference to how I envision the game to look and function. Think of it as like how you would make some UI in Photoshop and using that as a reference to when you implement it into the engine, this is basically a slightly more advanced way of that.